requirements.txt
selenium
pyautogui
pandas
python-dotenv